+++ src/client/Client.js
@@ -3,7 +3,7 @@
 const BaseClient = require('./BaseClient');
 const Permissions = require('../util/Permissions');
 const ClientVoiceManager = require('./voice/ClientVoiceManager');
-const WebSocketManager = require('./websocket/new/WebSocketManager');
+const WebSocketManager = require('./websocket/WebSocketManager');
 const ActionsManager = require('./actions/ActionsManager');
 const Collection = require('../util/Collection');
 const VoiceRegion = require('../structures/VoiceRegion');
@@ -16,8 +16,7 @@ const UserStore = require('../stores/UserStore');
 const ChannelStore = require('../stores/ChannelStore');
 const GuildStore = require('../stores/GuildStore');
 const GuildEmojiStore = require('../stores/GuildEmojiStore');
-const { Events, WSCodes, browser, DefaultOptions } = require('../util/Constants');
-const { delayFor } = require('../util/Util');
+const { Events, browser, DefaultOptions } = require('../util/Constants');
 const DataResolver = require('../util/DataResolver');
 const Structures = require('../util/Structures');
 const { Error, TypeError, RangeError } = require('../errors');
@@ -59,7 +58,7 @@ class Client extends BaseClient {
     }
 
     if (typeof this.options.shards === 'undefined' && this.options.shardCount) {
-      this.options.shards = Array.from({ length: this.options.shardCound }, (_, i) => i);
+      this.options.shards = Array.from({ length: this.options.shardCount }, (_, i) => i);
     }
 
     if (typeof this.options.shards === 'number') this.options.shards = [this.options.shards];
@@ -218,60 +217,6 @@ class Client extends BaseClient {
    * client.login('my token');
    */
   async login(token = this.token) {
-    if (!token || typeof token !== 'string') throw new Error('TOKEN_INVALID');
-    this.token = token = token.replace(/^(Bot|Bearer)\s*/i, '');
-    this.emit(Events.DEBUG, `Authenticating using token ${token}`);
-    let endpoint = this.api.gateway;
-    if (this.options.shardCount === 'auto') endpoint = endpoint.bot;
-    const res = await endpoint.get();
-    if (this.options.presence) {
-      this.options.ws.presence = await this.presence._parse(this.options.presence);
-    }
-    if (res.session_start_limit && res.session_start_limit.remaining === 0) {
-      const { session_start_limit: { reset_after } } = res;
-      this.emit(Events.DEBUG, `Exceeded identify threshold, setting a timeout for ${reset_after} ms`);
-      await delayFor(reset_after);
-    }
-    const gateway = `${res.url}/`;
-    if (this.options.shardCount === 'auto') {
-      this.emit(Events.DEBUG, `Using recommended shard count ${res.shards}`);
-      this.options.shardCount = res.shards;
-      this.options.totalShardCount = res.shards;
-      if (typeof this.options.shards === 'undefined' || !this.options.shards.length) {
-        this.options.shards = [];
-        for (let i = 0; i < this.options.shardCount; ++i) this.options.shards.push(i);
-      }
-    }
-    this.emit(Events.DEBUG, `Using gateway ${gateway}`);
-    this.ws.connect(gateway);
-    await new Promise((resolve, reject) => {
-      const onready = () => {
-        clearTimeout(timeout);
-        this.removeListener(Events.DISCONNECT, ondisconnect);
-        resolve();
-      };
-      const ondisconnect = event => {
-        clearTimeout(timeout);
-        this.removeListener(Events.READY, onready);
-        this.destroy();
-        if (WSCodes[event.code]) {
-          reject(new Error(WSCodes[event.code]));
-        }
-      };
-      const timeout = setTimeout(() => {
-        this.removeListener(Events.READY, onready);
-        this.removeListener(Events.DISCONNECT, ondisconnect);
-        this.destroy();
-        reject(new Error('WS_CONNECTION_TIMEOUT'));
-      }, this.options.shardCount * 25e3);
-      if (timeout.unref !== undefined) timeout.unref();
-      this.once(Events.READY, onready);
-      this.once(Events.DISCONNECT, ondisconnect);
-    });
-    return token;
-  }
-
-  async newLogin(token = this.token) {
     if (!token || typeof token !== 'string') throw new Error('TOKEN_INVALID');
     this.token = token = token.replace(/^(Bot|Bearer)\s*/i, '');
     this.emit(Events.DEBUG, `Provided token: ${token}`);
